refactor: removed agent/log.js and move the fn to agent connection
